[FEAT] 도장 다이얼로그 추가
