+++ app/src/main/java/com/polzzak_android/presentation/component/dialog/CommonDialogModel.kt
@@ -15,18 +15,27 @@ data class CommonDialogModel(
             DialogStyleType.ALERT -> {
                 require(content.calendar == null) { "Calendar must be null for ALERT type" }
                 require(content.mission == null) { "Mission must be null for ALERT type" }
+                require(content.stampImg == null) { "Stamp image must be null for ALERT type" }
             }
             DialogStyleType.CALENDAR -> {
                 require(content.calendar != null) { "Calendar must not be null for CALENDAR type" }
                 require(content.mission == null) { "Mission must be null for CALENDAR type" }
+                require(content.stampImg == null) { "Stamp image must be null for CALENDAR type" }
             }
             DialogStyleType.MISSION -> {
                 require(content.calendar == null) { "Calendar must be null for MISSION type" }
                 require(content.mission != null) { "Mission must not be null for MISSION type" }
+                require(content.stampImg == null) { "Stamp image must be null for MISSION type" }
             }
             DialogStyleType.LOADING -> {
                 require(content.calendar == null) { "Calendar must be null for LOADING type" }
                 require(content.mission == null) { "Mission must be null for LOADING type" }
+                require(content.stampImg == null) { "Stamp image must be null for LOADING type" }
+            }
+            DialogStyleType.STAMP -> {
+                require(content.calendar == null) { "Calendar must be null for STAMP type" }
+                require(content.mission == null) { "Mission must be null for STAMP type" }
+                require(content.stampImg != null) { "Stamp image not be null for STAMP type" }
             }
         }
     }
@@ -36,7 +45,8 @@ data class CommonDialogContent(
     val title: Spannable,
     val body: Spannable? = null,
     val calendar: Calendar? = null,
-    val mission: CommonDialogMissionData? = null
+    val mission: CommonDialogMissionData? = null,
+    val stampImg: Int? = null
 )
 
 data class CommonDialogMissionData(
@@ -46,5 +56,5 @@ data class CommonDialogMissionData(
 )
 
 enum class DialogStyleType {
-    ALERT, CALENDAR, MISSION, LOADING
+    ALERT, CALENDAR, MISSION, LOADING, STAMP
 }
